IMAGE: Split the generic QuickTime dithering into a separate class
